% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_pi_vec.R
\name{calc_pi_vec}
\alias{calc_pi_vec}
\title{Calculate the validation/sampling probabilities for a two-phase design.}
\usage{
calc_pi_vec(Validated, dat, phI_Y, phI_X)
}
\arguments{
\item{Validated}{Columns with the validation indicator (can be name or numeric index)}
}
\value{
1 x 4 vector of the sampling probabilities, pi:
\item{pi00}{P(V=1|\code{phI_Y}=0, \code{phI_X}=0)}
\item{pi01}{P(V=1|\code{phI_Y}=0, \code{phI_X}=1)}
\item{pi10}{P(V=1|\code{phI_Y}=1, \code{phI_X}=0)}
\item{pi11}{P(V=1|\code{phI_Y}=1, \code{phI_X}=1)}
}
\description{
Calculate the validation/sampling probabilities for a two-phase design.
}
