% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_size.R
\name{grid_size}
\alias{grid_size}
\title{Calculate size of the audit grid}
\usage{
grid_size(
  delta,
  phi,
  num_strat,
  phI_strat,
  prev_grid_des = NULL,
  prev_delta = NULL
)
}
\arguments{
\item{delta}{Audit step size (in people).}

\item{phi}{Number of people to be allocated (after \code{min_n}).}

\item{num_strat}{Number of strata on which sampling is based. Currently handles \code{num_strat} = 2, 4, or 8.}

\item{prev_grid_des}{If grid > 1, the audit from the previous iteration that was optimal.}

\item{sample_on}{Columns with the Phase I variables (should be categorical) used for sampling strata (can be name or numeric index). Currently, sampling on up to 3 variables can be accommodated.}
}
\value{
A vector of integers.
}
\description{
Calculate size of the audit grid
}
