% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{info}
\alias{info}
\title{Information matrix for two-phase measurement error problems.}
\usage{
info(pi_probs, indiv_score, sample_on)
}
\arguments{
\item{pi_probs}{Dataframe of sampling probabilities, including \code{sample_on}.}

\item{indiv_score}{Matrix of score vectors for all parameters, \code{beta} and \code{eta}.}

\item{sample_on}{Columns with the Phase I variables (should be categorical) used for sampling strata (can be name or numeric index).}
}
\value{
Information matrix.
}
\description{
Information matrix for two-phase measurement error problems.
}
