% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/od_loglik.R
\name{od_loglik}
\alias{od_loglik}
\title{Observed-data log-likelihood for measurement error settings with errors in outcome + exposure}
\usage{
od_loglik(params, dat, Y_val, Y_unval, X_val, X_unval, Validated)
}
\arguments{
\item{Y_val}{Column with the validated outcome (can be name or numeric index).}

\item{Y_unval}{Column with the unvalidated outcome (can be name or numeric index).}

\item{X_val}{Column(s) with the validated predictors (can be name or numeric index).}

\item{X_unval}{Column(s) with the unvalidated predictors (can be name or numeric index).}

\item{Validated}{Columns with the validation indicator (can be name or numeric index).}
}
\value{
Scalar function value.
}
\description{
Observed-data log-likelihood for measurement error settings with errors in outcome + exposure
}
