% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/od_loglik_Xonly.R
\name{od_loglik_Xonly}
\alias{od_loglik_Xonly}
\title{Observed-data log-likelihood for measurement error settings with errors in exposure only}
\usage{
od_loglik_Xonly(params, dat, Y_val, X_val, X_unval, Validated)
}
\arguments{
\item{Y_val}{Column with the validated outcome (can be name or numeric index).}

\item{X_val}{Column(s) with the validated predictors (can be name or numeric index).}

\item{X_unval}{Column(s) with the unvalidated predictors (can be name or numeric index).}

\item{Validated}{Columns with the validation indicator (can be name or numeric index).}
}
\value{
Scalar function value.
}
\description{
Observed-data log-likelihood for measurement error settings with errors in exposure only
}
