% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/od_loglik_Yonly.R
\name{od_loglik_Yonly}
\alias{od_loglik_Yonly}
\title{Observed-data log-likelihood for measurement error settings with errors in outcome only}
\usage{
od_loglik_Yonly(params, dat, Y_val, Y_unval, X_val, Validated)
}
\arguments{
\item{Y_val}{Column with the validated outcome (can be name or numeric index).}

\item{Y_unval}{Column with the unvalidated outcome (can be name or numeric index).}

\item{X_val}{Column(s) with the validated predictors (can be name or numeric index).}

\item{Validated}{Columns with the validation indicator (can be name or numeric index).}
}
\value{
Scalar function value.
}
\description{
Observed-data log-likelihood for measurement error settings with errors in outcome only
}
