% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optMLE_grid_close.R
\name{optMLE_grid_close}
\alias{optMLE_grid_close}
\title{Adaptive grid search optimization to find the optimal design (optMLE) - allows for closed strata}
\usage{
optMLE_grid_close(
  phI,
  phII,
  phI_strat,
  min_n,
  sample_on,
  closed = NULL,
  closed_at = NULL,
  indiv_score,
  return_full_grid = FALSE,
  max_grid_size = 10000
)
}
\arguments{
\item{phI}{Phase I sample size.}

\item{phII}{Phase II sample size.}

\item{phI_strat}{Phase I stratum sample sizes, named list.}

\item{min_n}{Minimum stratum size to be sampled.}

\item{sample_on}{Columns with the Phase I variables (should be categorical) used for sampling strata (can be name or numeric index). Currently, sampling on up to 3 binary variables can be accommodated.}

\item{closed}{For multi-wave designs, a vector of names for strata that are "closed", meaning we do not wish to sample from them. Default is \code{NULL}.}

\item{closed_at}{For multi-wave designs, a vector of already sampled sizes for strata that are "closed" (must be the same length as \code{closed}). Default is \code{NULL}.}

\item{indiv_score}{Matrix of score vectors for all parameters.}

\item{return_full_grid}{If \code{TRUE}, all audits from all iterations of the grid search will be return. Default is \code{FALSE}.}

\item{max_grid_size}{Integer maxium for the largest grids that will be searched.}
}
\value{
\item{all_opt}{Optimal designs chosen in each iteration of the grid search.}
\item{min_var}{Value of the variance achieved by the optimal design in the last iteration.}
\item{min_var_design}{Optimal design in the last iteration.}
\item{findOptimal}{\code{TRUE}/\code{FALSE} for whether final iteration found an optimal design.}
\item{full_grid_search}{If \code{return_full_grid} = TRUE, a dataframe containing all grids from ann iterations.}
\item{message}{Result of the grid search, options include \code{"No valid grids"}, \code{"Singular information"}, \code{"Tie for minimum"}, \code{"Grid completed without finding minimum"}, \code{"Grid search successful"}.}
}
\description{
Adaptive grid search optimization to find the optimal design (optMLE) - allows for closed strata
}
