% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_logistic.R
\name{prob_logistic}
\alias{prob_logistic}
\title{Logistic regression model probability P(Y|mu)}
\usage{
prob_logistic(y, mu)
}
\arguments{
\item{y}{number or numeric vector input for value of the logistic outcome}

\item{mu}{value of the linear predictor}
}
\description{
Logistic regression model probability P(Y|mu)
}
