% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_deviance.R
\name{sample_deviance}
\alias{sample_deviance}
\title{Sample deviance sample audit based on naive model from Phase I variables.}
\usage{
sample_deviance(
  formula,
  family,
  dat,
  phI,
  phII,
  tails = "both",
  wave1_Validated = NULL
)
}
\arguments{
\item{formula}{Model formula used to calculate residuals, passed to \code{glm()} or \code{glm.nb()}.}

\item{family}{Type of model to be used to fit \code{formula}, passed to \code{glm()} or \code{glm.nb()}. Currently, the function accepts \code{family = "gaussian", "binomial", "poisson", "log-binomial",} or \code{"negbin"}.}

\item{dat}{Dataframe or matrix containing variables from \code{formula}.}

\item{phI}{Phase I sample size.}

\item{phII}{Phase II sample size.}

\item{tails}{Which tails to sample from. Default is \code{tails = "both"}, but \code{"lower"} and \code{"upper"} only are possible.}

\item{wave1_Validated}{(For use with multi-wave designs) A logical vector with the validation indicator from first wave of audits.}
}
\value{
A vector of length \code{phI} with validation indicators V = 1 if selected for Phase II and V = 0 otherwise.
}
\description{
Sample deviance sample audit based on naive model from Phase I variables.
}
