% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_ets.R
\name{sample_ets}
\alias{sample_ets}
\title{Extreme-tail sampling based on a single Phase I variable.}
\usage{
sample_ets(ets_dat, phI, phII, wave1_Validated = NULL)
}
\arguments{
\item{ets_dat}{Vector containing one variable on which to sample.}

\item{phI}{Phase I sample size.}

\item{phII}{Phase II sample size.}

\item{wave1_Validated}{(For use with multi-wave designs) A logical vector with the validation indicator from first wave of audits.}
}
\value{
A vector of length \code{phI} with validation indicators V = 1 if selected for Phase II and V = 0 otherwise.
}
\description{
Extreme-tail sampling based on a single Phase I variable.
}
