% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_pca.R
\name{sample_pca}
\alias{sample_pca}
\title{Sample ETS-PCA audit based on Phase I variables.}
\usage{
sample_pca(pca_dat, phI, phII, wave1_Validated = NULL)
}
\arguments{
\item{pca_dat}{Dataframe or matrix containing variables on which to sample, passed to \code{princomp()}. This calculation is done using the correlation matrix, so these data do not need to be supplied on the scale.}

\item{phI}{Phase I sample size.}

\item{phII}{Phase II sample size.}

\item{wave1_Validated}{(For use with multi-wave designs) A logical vector with the validation indicator from first wave of audits.}
}
\value{
A vector of length \code{phI} with validation indicators V = 1 if selected for Phase II and V = 0 otherwise.
}
\description{
Sample ETS-PCA audit based on Phase I variables.
}
