% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_resid.R
\name{sample_resid}
\alias{sample_resid}
\title{Sample residual sample audit based on naive model from Phase I variables.}
\usage{
sample_resid(formula, family, dat, phI, phII, wave1_Validated = NULL)
}
\arguments{
\item{formula}{Model formula used to calculate residuals, passed to \code{glm()}.}

\item{family}{Type of model to be used to fit \code{formula}, passed to \code{glm()}. Currently, the function accepts \code{family = "gaussian", "binomial",} or \code{"poisson"}.}

\item{dat}{Dataframe or matrix containing variables from \code{formula}.}

\item{phI}{Phase I sample size.}

\item{phII}{Phase II sample size.}

\item{wave1_Validated}{(For use with multi-wave designs) A logical vector with the validation indicator from first wave of audits.}
}
\value{
A vector of length \code{phI} with validation indicators V = 1 if selected for Phase II and V = 0 otherwise.
}
\description{
Sample residual sample audit based on naive model from Phase I variables.
}
