% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_sfs.R
\name{sample_sfs}
\alias{sample_sfs}
\title{Sample score function sample audit based on naive model from Phase I variables.}
\usage{
sample_sfs(formula, family, dat, phI, phII, X, absValue = FALSE)
}
\arguments{
\item{formula}{Model formula used to calculate residuals, passed to \code{glm()}.}

\item{family}{Type of model to be used to fit \code{formula}, passed to \code{glm()}. Currently, the function accepts \code{family = "gaussian", "binomial",} or \code{"poisson"}.}

\item{dat}{Dataframe or matrix containing variables from \code{formula}.}

\item{phI}{Phase I sample size.}

\item{phII}{Phase II sample size.}

\item{X}{Column with the error-prone covariate incorporated into sampling (can be name or numeric index).}

\item{absValue}{Logical, if \code{absValue = FALSE} (the default) the design is based on the weighted score functions, rather than their absolute values.}
}
\value{
A vector of length \code{phI} with validation indicators V = 1 if selected for Phase II and V = 0 otherwise.
}
\description{
Sample score function sample audit based on naive model from Phase I variables.
}
