% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_srs.R
\name{sample_srs}
\alias{sample_srs}
\title{Sample simple random sample (SRS) audit.}
\usage{
sample_srs(phI, phII, wave1_Validated = NULL)
}
\arguments{
\item{phI}{Phase I sample size.}

\item{phII}{Phase II sample size.}

\item{wave1_Validated}{(For use with multi-wave designs) A logical vector with the validation indicator from first wave of audits.}
}
\value{
A vector of length \code{phI} with validation indicators V = 1 if selected for Phase II and V = 0 otherwise.
}
\description{
Sample simple random sample (SRS) audit.
}
