% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_Xonly.R
\name{score_Xonly}
\alias{score_Xonly}
\title{Score vectors for measurement error settings with errors in exposure only}
\usage{
score_Xonly(comp_dat, Y_val, X_val, X_unval, Validated, beta, eta)
}
\arguments{
\item{comp_dat}{Dataframe with complete data.}

\item{Y_val}{Column with the validated outcome (can be name or numeric index).}

\item{X_val}{Column(s) with the validated predictors (can be name or numeric index).}

\item{X_unval}{Column(s) with the unvalidated predictors (can be name or numeric index).}

\item{Validated}{Columns with the validation indicator (can be name or numeric index).}

\item{beta}{Parameter value for the log odds ratio of \code{X_val} on \code{Y_val}.}

\item{eta}{Parameter values for other nuisance parameters.}
}
\value{
Matrix with one column for the score vector for each parameter \code{beta}, \code{eta}.
}
\description{
Score vectors for measurement error settings with errors in exposure only
}
