% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigmoid.R
\name{sigmoid}
\alias{sigmoid}
\title{Sigmoid function: f(x) = (1 + exp(x)) ^ (-1)}
\usage{
sigmoid(x)
}
\arguments{
\item{x}{Input number or numeric vector}
}
\description{
Sigmoid function: f(x) = (1 + exp(x)) ^ (-1)
}
