% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suggest_step.R
\name{suggest_step}
\alias{suggest_step}
\title{Recommends next audit step size for iterative grid search.}
\usage{
suggest_step(
  phII,
  phI_strat,
  min_n,
  num_strat,
  num_steps = "smallest",
  prev_grid_des,
  prev_delta,
  max_grid_size
)
}
\arguments{
\item{phII}{Phase II sample size.}

\item{min_n}{Minimum stratum size to be sampled.}

\item{num_strat}{Number of strata on which sampling is based.}

\item{num_steps}{Character option for how many steps to be returned, options include \code{"smallest"} and \code{"all"}. DEFAULT is \code{"smallest"}.}

\item{prev_grid_des}{If grid > 1, the audit from the previous iteration that was optimal.}

\item{max_grid_size}{Integer maximum for the largest grids that will be searched.}

\item{prev_grid_delta}{If grid > 1, the step size from the previous iteration.}
}
\value{
An integer.
}
\description{
Recommends next audit step size for iterative grid search.
}
