% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twophase_mle_int.R
\name{twophase_mle_interact}
\alias{twophase_mle_interact}
\alias{twophase_mle_int}
\title{Maximum likelihood estimator for logistic regression with errors in binary outcome + exposure (with interaction)}
\usage{
twophase_mle_int(
  dat,
  Y_val,
  Y_unval = NULL,
  X_val,
  X_unval = NULL,
  addl_covar = NULL,
  interact = NULL,
  Validated,
  int_Y_unval = FALSE,
  int_X_unval = FALSE,
  noSE = FALSE
)
}
\arguments{
\item{dat}{Data containing all columns referenced in \code{Y_unval}, \code{Y_val}, \code{X_unval}, \code{X_val}, \code{addl_covar}, and \code{Validated}.}

\item{Y_val}{Column with the validated outcome (can be name or numeric index).}

\item{Y_unval}{Column with the unvalidated outcome (can be name or numeric index). If outcome is error-free, \code{Y_unval = NULL} (DEFAULT).}

\item{X_val}{Column(s) with the validated predictors (can be name or numeric index).}

\item{X_unval}{Column(s) with the unvalidated predictors (can be name or numeric index). If covariates are error-free, \code{X_unval = NULL} (DEFAULT).}

\item{addl_covar}{Column(s) with additional error-free covariates (can be name or numeric index). Default is \code{NULL}.}

\item{interact}{Column(s) with interaction term(s) (can be name or numeric index). Default is \code{NULL}.}

\item{Validated}{Columns with the validation indicator (can be name or numeric index).}

\item{int_Y_unval}{If \code{TRUE}, misclassification model for \code{Y_unval} includes \code{interact}. Default is \code{FALSE}.}

\item{int_X_unval}{If \code{TRUE}, misclassification model for \code{X_unval} includes \code{interact}. Default is \code{FALSE}.}

\item{noSE}{If \code{TRUE}, standard errors are calculated by inverting the hessian at convergence. Default is \code{FALSE}.}
}
\value{
List of dataframes with model coefficients.
\item{mod_Y_unval}{Outcome misclassification model coefficients}
\item{mod_X_unval}{Exposure misclassification model coefficients}
\item{mod_Y_val}{Analysis model coefficients}
\item{mod_X_val}{Exposure model coefficients}
\item{conv}{\code{TRUE}/\code{FALSE} for convergence}
}
\description{
Maximum likelihood estimator for logistic regression with errors in binary outcome + exposure (with interaction)
}
