% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_grid.R
\name{visualize_grid}
\alias{visualize_grid}
\title{Creates a graphic representation of the input adaptive grid search}
\usage{
visualize_grid(grid_search)
}
\arguments{
\item{grid_search}{\code{List} object returned from \code{optMLE_grid()} with option \code{return_full_grid = TRUE}.}
}
\description{
Creates a graphic representation of the input adaptive grid search
}
