% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_grid.R, R/grid_size.R
\name{grid_size}
\alias{grid_size}
\alias{build_grid}
\title{Build audit grid}
\usage{
build_grid(
  delta,
  phi,
  num_strat,
  phI_strat,
  phIIa_strat = NULL,
  min_n,
  prev_grid_des = NULL,
  prev_delta = NULL,
  grid_window = NULL
)

grid_size(
  delta,
  phi,
  num_strat,
  phI_strat,
  prev_grid_des = NULL,
  prev_delta = NULL
)
}
\arguments{
\item{delta}{Audit step size (in people).}

\item{phi}{Number of people to be allocated.}

\item{num_strat}{Number of strata on which sampling is based.}

\item{phI_strat}{Phase I stratum sample sizes, named list.}

\item{phIIa_strat}{For multi-wave designs, Phase II(a) stratum sample sizes as a named list, dataframe, or vector. Default is \code{NULL}.}

\item{min_n}{Minimum stratum size to be sampled.}

\item{prev_grid_des}{If grid > 1, the audit from the previous iteration that was optimal.}

\item{prev_delta}{If grid > 1, the step size from the previous iteration.}

\item{grid_window}{As an alternative to \code{prev_delta}, specify a custom \code{grid_window = c(start, end)} around \code{prev_grid_des}.}
}
\value{
An integer.
}
\description{
Build audit grid

Calculate size of the audit grid
}
