% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_grid.R, R/grid_size.R
\name{grid_size}
\alias{grid_size}
\alias{build_grid}
\title{Build audit grid}
\usage{
build_grid(
  delta,
  phi,
  num_strat,
  phI_strat,
  phIIa_strat = NULL,
  closed = NULL,
  closed_at = NULL,
  min_n,
  prev_grid_des = NULL,
  prev_delta = NULL
)

grid_size(
  delta,
  phi,
  num_strat,
  phI_strat,
  closed = NULL,
  prev_grid_des = NULL,
  prev_delta = NULL
)
}
\arguments{
\item{delta}{Audit step size (in people).}

\item{phi}{Number of people to be allocated: \code{phi = phII - num_strat * min_n}.}

\item{num_strat}{Number of strata on which sampling is based. Currently handles \code{num_strat} = 2, 4, or 8.}

\item{phI_strat}{Phase I stratum sample sizes, named list.}

\item{phIIa_strat}{For multi-wave designs, Phase II(a) stratum sample sizes as a named list, dataframe, or vector. Default is \code{NULL}.}

\item{closed}{For multi-wave designs, a vector of names for strata that are "closed", meaning we do not wish to sample from them. Default is \code{NULL}.}

\item{closed_at}{For multi-wave designs, a vector of already sampled sizes for strata that are "closed" (must be the same length as \code{closed}). Default is \code{NULL}.}

\item{min_n}{Minimum stratum size to be sampled.}

\item{prev_grid_des}{If grid > 1, the audit from the previous iteration that was optimal.}

\item{prev_delta}{If grid > 1, the step size from the previous iteration.}
}
\value{
An integer.
}
\description{
Build audit grid

Calculate size of the audit grid
}
