% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optMLE_grid.R
\name{optMLE_grid}
\alias{optMLE_grid}
\title{Adaptive grid search optimization to find the optimal design (optMLE)}
\usage{
optMLE_grid(
  phI,
  phII,
  phI_strat,
  min_n,
  sample_on,
  indiv_score,
  return_full_grid = FALSE,
  max_grid_size = 10000
)
}
\arguments{
\item{phI}{Phase I sample size.}

\item{phII}{Phase II sample size.}

\item{phI_strat}{Phase I stratum sample sizes, named list.}

\item{min_n}{Minimum stratum size to be sampled.}

\item{sample_on}{Columns with the Phase I variables (should be categorical) used for sampling strata (can be name or numeric index). Currently, sampling on up to 3 binary variables can be accommodated.}

\item{indiv_score}{Matrix of score vectors for all parameters, \code{beta} and \code{eta}.}

\item{return_full_grid}{If \code{TRUE}, all audits from all iterations of the grid search will be return. Default is \code{FALSE}.}
}
\value{
List
}
\description{
Adaptive grid search optimization to find the optimal design (optMLE)
}
