% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suggest_step.R
\name{suggest_step}
\alias{suggest_step}
\title{Recommends next audit step size for iterative grid search.}
\usage{
suggest_step(
  phII,
  phI_strat,
  min_n,
  num_strat,
  closed,
  prev_grid_des,
  prev_delta,
  max_grid_size
)
}
\arguments{
\item{phII}{Phase II sample size.}

\item{min_n}{Minimum stratum size to be sampled.}

\item{num_strat}{Number of strata on which sampling is based. Currently handles \code{num_strat} = 2, 4, or 8.}

\item{closed}{For multi-wave designs, a vector of names for strata that are "closed", meaning we do not wish to sample from them. Default is \code{NULL}.}

\item{prev_grid_des}{If grid > 1, the audit from the previous iteration that was optimal.}

\item{max_grid_size}{Integer maxium for the largest grids that will be searched.}

\item{prev_grid_delta}{If grid > 1, the step size from the previous iteration.}
}
\value{
An integer.
}
\description{
Recommends next audit step size for iterative grid search.
}
