% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suggest_step.R
\name{suggest_step}
\alias{suggest_step}
\title{Recommends next audit step size for iterative grid search.}
\usage{
suggest_step(
  phII,
  phI_strat,
  min_n,
  sample_on,
  prev_grid_des = NULL,
  prev_delta = NULL,
  max_grid_size = 25000
)
}
\arguments{
\item{phII}{Phase II sample size.}

\item{min_n}{Minimum stratum size to be sampled.}

\item{sample_on}{Columns with the Phase I variables (should be categorical) used for sampling strata (can be name or numeric index). Currently, sampling on up to 3 variables can be accommodated.}

\item{prev_grid_des}{If grid > 1, the audit from the previous iteration that was optimal.}

\item{max_grid_size}{Integer maxium for the largest grids that will be searched. DEFAULT = 100000.}

\item{prev_grid_delta}{If grid > 1, the step size from the previous iteration.}
}
\value{
An integer.
}
\description{
Recommends next audit step size for iterative grid search.
}
