% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_formula.R
\name{var_formula}
\alias{var_formula}
\title{Derived asymptotic variance of the MLE for beta, the logOR of X on Y}
\usage{
var_formula(pi_vec, phI, indiv_score, sample_on)
}
\arguments{
\item{pi_vec}{Vector of sampling probabilities (returned from \code{calc_pi_vec}).}

\item{phI}{Phase I sample size.}

\item{indiv_score}{Matrix of score vectors for all parameters, \code{beta} and \code{eta} (returned from \code{score}).}

\item{sample_on}{Columns with the Phase I variables (should be categorical) used for sampling strata (can be name or numeric index). Currently, sampling on up to 3 variables can be accommodated.}
}
\value{
Scalar function value for Var(beta).
}
\description{
Derived asymptotic variance of the MLE for beta, the logOR of X on Y
}
